refactor: split the graphical renderer into a module directory
`src/handlers/graphical.rs` had grown to 1624 lines mixing configuration,
document-level rendering, source-snippet drawing, and the supporting
`Line`/`FancySpan` types. Split it into `src/handlers/graphical/` with one
responsibility per file:

- handler.rs — the `GraphicalReportHandler` struct + builders (public API)
- report.rs  — diagnostic-level rendering: title, causes, help/note, related
- snippet.rs — read & merge labelled spans, render each context
- gutter.rs  — line numbers + multi-line span gutters
- label.rs   — underlines and labels drawn under the source text
- line.rs    — the `Line` model: line splitting, span geometry, width math
- span.rs    — `FancySpan` + `LabelRenderMode`

Every hot-path function is moved verbatim, so the split is codegen-neutral.
The only logic changes are two readability cleanups on cold/neutral paths:
deduplicate the 4x-repeated `textwrap::Options` builder into a `wrap_options`
helper, and lift `CharWidthIterator` out of its host method to module scope.

Rendered output is unchanged — the snapshot suite and the span-reading
differential fuzz test pass byte-for-byte identically.
